feat: Support insect analysis and add insect icon
The application now supports analyzing images of insects as well as plants. The Gemini API prompt has been updated to include insect analysis, and the `AnalysisResponse` and `AnalysisResultData` types now include an `isInsect` field.

UI text and the `BugIcon` component have been added to reflect this new functionality. History items will now display a bug icon if the analysis was of an insect.